Updated polynomial interpolator.

+++ code/engine/vipredictors/vifourierpredictor/headers/vifourierpredictor.h
@@ -72,7 +72,7 @@ class ViFourierPredictor : public ViPredictor
 		Estimation mEstimation;
 		Mode mMode;
 
-		QVector<QVector<int>> mBestParameters;
+		QVector<QVector<qint64>> mBestParameters;
 
 		int mDegree;
 		int mDerivatives;
